use lazy evaluation for configuration attributes
This change allows override attributes to be defined in recipes (during the compile phase) of wrapper cookbooks
and still take effect before the resource is converged (aka executed).

Without this change, it is dangerous to set overrides in recipes because the order in which recipes are run
determines whether the overrides will actually take effect when the resources are converged.